Adding font awesome icons and creating a settings component
