+++ client/package.json
@@ -10,6 +10,9 @@
   },
   "dependencies": {
     "@emotion/react": "^11.10.5",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@mantine/core": "^5.9.6",
     "@mantine/hooks": "^5.9.6",
     "@react-three/drei": "^9.48.1",
